Crate yogcrypt

A fast, general purpose crypto library in pure Rust.

YogCrypt is designed to be a high-performance, general purpose crypto library.

YogCrypt currently provides three cryptographic algorithms in Chinese National Standard, namely the SM2 cryptographic asymmetric algorithm, the SM3 cryptographic hash algorithm, and the SM4 block cipher algorithm.

License

YogCrypt is distributed under the terms of both the MIT license and the Apache License (Version 2.0). See LICENSE-APACHE and LICENSE-MIT for details.

Modules

sm2
An implementation of the SM2 signature standard.

sm3
An implementation of the SM3 cryptographic hash standard.

sm4
An implementation of the SM4 block cipher standard.